+++ src/pkg/images/pull.go
@@ -148,7 +148,7 @@ func Pull(ctx context.Context, imageList []transform.Image, destinationDirectory
 		}
 	}
 
-	l.Debug("gathering credentials from default Docker config file", "credentials_configured", credStore.IsAuthConfigured())
+	l.Debug("gathering credentials from default Docker config file", "credentialsConfigured", credStore.IsAuthConfigured())
 	platform := &ocispec.Platform{
 		Architecture: opts.Arch,
 		// TODO: in the future we could support Windows images
@@ -491,7 +491,7 @@ func orasSave(ctx context.Context, imageInfo imagePullInfo, opts PullOptions, ds
 			if config.ZarfDefaultRetries > 1 && n+1 < uint(config.ZarfDefaultRetries) {
 				l.Warn("retrying image pull",
 					"attempt", n+1,
-					"max_attempts", config.ZarfDefaultRetries,
+					"maxAttempts", config.ZarfDefaultRetries,
 					"image", imageInfo.registryOverrideRef,
 					"error", err,
 				)